Fix/adjust pass submission;
When lovr.graphics.submit is called, it invalidates all of the passes
that were created that frame.  Generally this is supposed to be called
once or twice each frame to submit all of the passes, since submission
is not a lightweight call.

Currently GenerateOSKTextures individually submits 1 pass for each
keyboard mode, which will invalidate the other passes that are about
to be submitted (i.e. main_pass).  This patch changes it to instead
return a table of passes that gets added to the list of passes that
RenderFrame submits.

It may also be good to do this same thing to RenderFrame, where instead
of submitting the passes it just returns them to the caller.  However, I
didn't do that here since it would be a change to the external API and
probably needs more discussion.

The demo also returns true from lovr.draw so that the main pass doesn't
get submitted twice.  Annoyingly, LÖVR doesn't complain about this right
now (fix incoming).